LF-3911 rename slide menu to side menu
