+++ packages/webapp/src/components/Navigation/index.jsx
@@ -3,7 +3,7 @@ import { useTheme } from '@mui/styles';
 import { IconButton, useMediaQuery } from '@mui/material';
 import { Menu } from '@mui/icons-material';
 import TopMenu from './Menus/TopMenu';
-import SlideMenu from '../../containers/Navigation/SlideMenu';
+import SlideMenu from '../../containers/Navigation/SideMenu';
 import {
   NavbarSpotlightProvider,
   NavBarNotificationSpotlightProvider,
